requirements.txt
alembic==1.7.5
attrs==21.2.0
factory-boy==3.2.1
Faker==9.8.0
greenlet==1.1.2
iniconfig==1.1.1
Mako==1.1.5
MarkupSafe==2.0.1
packaging==21.2
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.11.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==1.4.27
text-unidecode==1.3
toml==0.10.2
